Uses classes prototype instead of their aliases
 - Class chart was moved to its own script.
 - Class AxisInternal was moved to its own script. It's only needed by Class Axis
 - Definition of Chart removed from core.js
 - Definition of AxisInternal removed from axis.js
