+++ src/util.js
@@ -1,4 +1,5 @@
 import {ChartInternal} from './chart-internal';
+import {Chart} from './chart';
 
 export var c3 = { version: "0.6.3" ,
     chart: {
@@ -6,31 +7,12 @@ export var c3 = { version: "0.6.3" ,
         internal: {
             fn: ChartInternal.prototype,
         }
+    },
+    generate: function (config) {
+      return new Chart(config);
     }
 };
 
-export function Chart(config) {
-    var $$ = this.internal = new ChartInternal(this);
-    $$.loadConfig(config);
-
-    $$.beforeInit(config);
-    $$.init();
-    $$.afterInit(config);
-
-    // bind "this" to nested API
-    (function bindThis(fn, target, argThis) {
-        Object.keys(fn).forEach(function (key) {
-            target[key] = fn[key].bind(argThis);
-            if (Object.keys(fn[key]).length > 0) {
-                bindThis(fn[key], target[key], argThis);
-            }
-        });
-    })(c3.chart.fn, this, this);
-}
-
-c3.generate = function (config) {
-    return new Chart(config);
-};
 
 export function Component(owner, componentKey, fn) {
     this.owner = owner;
@@ -68,7 +50,7 @@ export var isEmpty = function (o) {
     return typeof o === 'undefined' || o === null || (isString(o) && o.length === 0) || (typeof o === 'object' && Object.keys(o).length === 0);
 };
 export var notEmpty = function (o) {
-    return !c3.chart.internal.fn.isEmpty(o);
+    return !ChartInternal.prototype.isEmpty(o);
 };
 export var getOption = function (options, key, defaultValue) {
     return isDefined(options[key]) ? options[key] : defaultValue;
